Removes logging and migrates module/kusto/vendor to module/vendor
Removed logging in all locations except for unsafe.  We allow that logging message to be suppressed.

One logging line was commented out instead of removed, that is done on purpose.

Moved the vendor folder to the root, which we should have done when we moved the go.mod file to the root.