Remove header pool and Accept-Encoding deflate, 2x performance improvement (#19)
Starting with the header pool.  This seems like some premature optimization that is going to have little affect for this package.  Removing this until such a time that this is a bottleneck in some way, which is doubtful.

Removed Accept-Endoing "deflate".  I'm not actually sure what is happening here.

Deflate should use the deflate compression with zlib headers.
Gzip should use deflate compression with gzip headers.

This should mean that performance wise we get zero difference.  However, that is not the caese.

If we use Accept-Encoding: "gzip, deflate" , it chooses deflate.  Deflate has a 2x cost in speed from the backend.

I suspsect that the backend's implementation of deflate is bad for some reason.

Removing deflate forces the backend to use gzip and doubled our download performance.  This puts the Go performance and .Net performance at the same rate, which seems limited by either the network or the speed of the backend.

Co-authored-by: John Doak <jdoak@janama-2.redmond.corp.microsoft.com>