Convert Kusto Dynamic.Value to []byte type from map[string]interface{}
The problem is that we need to support []interface{} and we don't want to force the user to Marshal data more than once.

This costs us a little, as the json library puts our data into a map[string]string , []interface{} or interface{}.

But I think strategically this makes sense for the user.  They can decide how they want to unmarshal it.