Convert Kusto Dynamic.Value to []byte type from map[string]interface{} (#28)
* Refactor decoding and parallelizing tests

Decoding was being done into a map[string]interface{}, then manually decoded into our structs.  This made us do a lot of boilerplate and validation of fields we shouldn't have to worry about.

This moves us to a lower allocation type of decoding and significantly less code.  But it did require forking the standard library's encoding/json package in order to work.

There was a lot of experimentation here to figure an optimal way to do this.  I also tried third party packages, but they don't like streaming or require code generation which seemed heavy handed.  I made two line changes to json which solved all our issues.  This is not portable back to the standard library, because it would violate the 1.0 compatibility promise.  It is recommended by the go devs to fork if you need these kinds of fixes.  There is a readme included that tells the changes.

Also made all the tests parallel, which takes already fast tests and makes them even faster.  Basically all user time is spent waiting for the end to end test.

* Removed some byte values that were not needed.

* Convert Kusto Dynamic.Value to []byte type from map[string]interface{}

The problem is that we need to support []interface{} and we don't want to force the user to Marshal data more than once.

This costs us a little, as the json library puts our data into a map[string]string , []interface{} or interface{}.

But I think strategically this makes sense for the user.  They can decide how they want to unmarshal it.

Co-authored-by: John Doak <jdoak@janama-2.redmond.corp.microsoft.com>
Co-authored-by: John Doak <jdoak@Fan061719.northamerica.corp.microsoft.com>